+++ src/Application/Command/Sheet/Template/SavePrintTemplate.php
@@ -1,18 +1,11 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
-
 namespace Proximum\Vimeet\Application\Command\Sheet\Template;
 
+use Proximum\Vimeet\Application\Command\Command;
 use Proximum\Vimeet\Domain\Model\Template\SheetTemplate;
 
-class SavePrintTemplate
+class SavePrintTemplate implements Command
 {
     /** @var SheetTemplate */
     public $template;
@@ -20,10 +13,6 @@ class SavePrintTemplate
     /** @var array */
     public $value;
 
-    /**
-     * @param SheetTemplate $template
-     * @param array         $value
-     */
     public function __construct(SheetTemplate $template, array $value)
     {
         $this->template = $template;